finish benchmarking
